// admin/src/generated/api/audit/models/auditLogDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */
import type { AuditLogDtoBefore } from './auditLogDtoBefore';
import type { AuditLogDtoAfter } from './auditLogDtoAfter';

export interface AuditLogDto {
  id: string;
  requestId: string;
  sequenceNo: number;
  actorType: string;
  /** @nullable */
  actorUserId?: string | null;
  /** @nullable */
  actorDisplayName?: string | null;
  action: string;
  entityType: string;
  /** @nullable */
  entityId?: string | null;
  /** @nullable */
  before?: AuditLogDtoBefore;
  /** @nullable */
  after?: AuditLogDtoAfter;
  /** @nullable */
  reason?: string | null;
  createdAt: string;
}